Remove ThemeContextProvider
