+++ src/context/ThemeContextProvider.js
@@ -1,0 +1,29 @@
+import { createContext, useEffect, useState } from "react";
+
+export const ThemeContext = createContext();
+
+const ThemeContextProvider = ({ children }) => {
+  const [theme, setTheme] = useState("light");
+
+  useEffect(() => {
+    const time = new Date().getHours();
+
+    if (time >= 6 && time <= 16) {
+      setTheme("light");
+    } else {
+      setTheme("dark");
+    }
+  }, []);
+
+  const toggleTheme = () => {
+    theme === "light" ? setTheme("dark") : setTheme("light");
+  };
+
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
+
+export default ThemeContextProvider;
