--- src/context/ThemeProvider.js
@@ -1,23 +1,0 @@
-import { createContext, useEffect, useState } from "react";
-
-export const ThemeContext = createContext();
-
-export const ThemeProvider = ({ children }) => {
-  const [isDark, setIsDark] = useState(false);
-  const theme = isDark ? "_dark" : "";
-  const toggleTheme = () => {
-    localStorage.setItem("isDark", JSON.stringify(!isDark));
-    setIsDark(!isDark);
-  };
-
-  useEffect(() => {
-    const isDark = localStorage.getItem("isDark") === "true";
-    setIsDark(isDark);
-  }, []);
-
-  return (
-    <ThemeContext.Provider value={[{ theme, isDark }, toggleTheme]}>
-      {children}
-    </ThemeContext.Provider>
-  );
-};
